Decide type of plain BSON array based on its contents
If an array has no Julia type information, then we previously made it
Vector{Any}. Even if it only contained a single concrete data type. Now we
create an array from the firts element's type and expand the type if another
type of object is discovered during parsing.

This may be surprising behaviour to some people, but this is also generally
how Julia works if you create an array from an iterator or untyped
literal. Certainly it could break some peoples code so I will bump the version
to 0.6.0.

If you really need an array of type Any after deserialisation, then just do
something like `Any[A...]`.

https://discourse.julialang.org/t/tail-call-optimization-and-function-barrier-based-accumulation-in-loops/25831

This doesn't happen when the array is referenced because we would have to
update the reference after type promotion and it is not clear if this is worth
the effort to implement.